imp setTime

+++ react-app/src/components/viewTasks.js
@@ -9,10 +9,17 @@ function ViewTask(props){
             return list.filter(l => l.id !== id)
     }
     
-    const deleteTask = (id_task) =>{
-        let listOfTasks = props.tasks;
-        listOfTasks = deleteForId(id_task, listOfTasks)
-        props.updatingListOfTasks(listOfTasks)
+    const deleteTask = async(id_task) =>{
+        try{
+            const res = await axios.delete("http://localhost:3001/tasks/delete/"+id_task, {headers:{'Authorization':`Bearer ${props.token}`}});
+            if(res.data.message === "Task deleted"){
+                let listOfTasks = props.tasks;
+                listOfTasks = deleteForId(id_task, listOfTasks)
+                props.updatingListOfTasks(listOfTasks)
+            }
+        }catch(err){
+            console.log(err);
+        }    
     }
 
     const editCheckTask = (id_task) =>{
@@ -22,20 +29,28 @@ function ViewTask(props){
         props.updatingListOfTasks(listOfTasks)
     }
 
-    const editOrAddTask = async (task, isEditTask) =>{
+    const editOrAddTask = async (task, taskSv,isEditTask) =>{
         const urlAdd ="http://localhost:3001/tasks/new"
         
-        let listOfTasks = props.tasks; //ojo aca, no se si un componente lo acomoda........ojo con el add
+        let listOfTasks = props.tasks;
         if(isEditTask){
-            const index = listOfTasks.findIndex(t => t.id === task.id);
-            listOfTasks[index] = task;
-            props.updatingListOfTasks(listOfTasks);
+            try {
+                console.log("entre a la edicion")
+                const res = await axios.patch("http://localhost:3001/tasks/update/"+task.id, taskSv, {headers:{'Authorization':`Bearer ${props.token}`}});
+                if(res.data.message === "Task updated"){
+                    const index = listOfTasks.findIndex(t => t.id === task.id);
+                    listOfTasks[index] = task;
+                    props.updatingListOfTasks(listOfTasks);
+                }
+            } catch (error) {
+                console.log(error);
+            }
+            
         }else{
             
             try {
                 const res = await axios.post(urlAdd,{"title":task.title}, {headers:{'Authorization':`Bearer ${props.token}`}});
                 if (res.data.message === "Task created") {
-                    console.log("task creada")
                     listOfTasks.push(task)
                     props.updatingListOfTasks(listOfTasks)
                 }
